+++ domain/src/main/java/pro/number/domain/model/Group.kt
@@ -11,14 +11,14 @@ data class Group(
     val itemsCount: Int = DEFAULT_ITEM_COUNT,
     val tips: Byte = DEFAULT_TIPS,
     val waiter: String = DEFAULT_WAITER,
-    val total: Int = DEFAULT_TOTAL
+    val total: Float = DEFAULT_TOTAL
 ) {
     companion object {
         private const val UNDEFINED_ID = 0L
         private const val BASE_NAME = "Новая группа"
         private const val DEFAULT_ITEM_COUNT = 0
         private const val DEFAULT_TIPS = 0.toByte()
-        private const val DEFAULT_TOTAL = 0
+        private const val DEFAULT_TOTAL = 0f
         private const val DATE_PATTERN = "yyyy-MM-dd"
         private const val DEFAULT_WAITER = ""
     }
